use constructor overloads
